Move packaging options to setup.cfg
Using setup.cfg offers a few benefits with the INI file syntax over
using the options in the setup.py file. Files can be directly
specified for the description, instead of read by the script.
The format is more declarative than imperative, since it is not part
of the script. Also add pyproject.toml to add PEP 517 compatibility.