Update conda recipe and fix tests and description
Bump dependency versions, verified to work correctly. Fix how tests
are called after a build, due to the change to the src directory
style layout, and no longer distributing the tests with the
package. Fix the long description, which caused parsing errors for
the not-quite-the-same-as-GitHub Markdown parser that Anaconda.org
uses.