Set up Matplotlib for use with Pint
Pint now includes a function to use Matplotlib correctly. This requires
Pint >=0.9. With this function, Matplotlib takes arrays with units
attached and plots them automatically. This also requires small changes
to the examples. np.arange does not support Quantities as arguments, so
switch to np.linspace which has that support. Create arrays for
accumulation of results, since lists don't seem to work properly with
when units are involved.