Build CoolProp so we can test on Python 3.8
CoolProp 6.3.0 from PyPI raises a TypeError due to some changes that
since the release of Python 3.8. The master branch of CoolProp has
the fixes, so build that and test ThermoState on Python 3.8.

Don't do this on Windows, since Windows should be covered by the
other Python versions, and I can't easily test compiling CoolProp
on Windows right now.